+++ capacity-checker-sdk/src/main/java/com/dwarfeng/capacitychecker/sdk/bean/entity/FastJsonAlarmInfo.java
@@ -1,0 +1,138 @@
+package com.dwarfeng.capacitychecker.sdk.bean.entity;
+
+import com.alibaba.fastjson.annotation.JSONField;
+import com.dwarfeng.capacitychecker.stack.bean.entity.AlarmInfo;
+import com.dwarfeng.subgrade.sdk.bean.key.FastJsonLongIdKey;
+import com.dwarfeng.subgrade.stack.bean.Bean;
+
+import java.util.Date;
+import java.util.Objects;
+
+/**
+ * FastJson 报警信息。
+ *
+ * @author DwArFeng
+ * @since 1.0.0
+ */
+public class FastJsonAlarmInfo implements Bean {
+
+    private static final long serialVersionUID = -1370594829639018928L;
+
+    public static FastJsonAlarmInfo of(AlarmInfo alarmInfo) {
+        if (Objects.isNull(alarmInfo)) {
+            return null;
+        } else {
+            return new FastJsonAlarmInfo(
+                    FastJsonLongIdKey.of(alarmInfo.getKey()), alarmInfo.getLimitCapacity(),
+                    alarmInfo.getActualCapacity(), alarmInfo.getRatio(), alarmInfo.getHappenedDate(),
+                    alarmInfo.getAlarmMessage(), alarmInfo.isAlarming()
+            );
+        }
+    }
+
+    @JSONField(name = "key", ordinal = 1)
+    private FastJsonLongIdKey key;
+
+    @JSONField(name = "limit_capacity", ordinal = 2)
+    private long limitCapacity;
+
+    @JSONField(name = "actual_capacity", ordinal = 3)
+    private long actualCapacity;
+
+    @JSONField(name = "ratio", ordinal = 4)
+    private double ratio;
+
+    @JSONField(name = "happened_date", ordinal = 5)
+    private Date happenedDate;
+
+    @JSONField(name = "alarm_message", ordinal = 6)
+    private String alarmMessage;
+
+    @JSONField(name = "alarming", ordinal = 7)
+    private boolean alarming;
+
+    public FastJsonAlarmInfo() {
+    }
+
+    public FastJsonAlarmInfo(
+            FastJsonLongIdKey key, long limitCapacity, long actualCapacity, double ratio, Date happenedDate,
+            String alarmMessage, boolean alarming
+    ) {
+        this.key = key;
+        this.limitCapacity = limitCapacity;
+        this.actualCapacity = actualCapacity;
+        this.ratio = ratio;
+        this.happenedDate = happenedDate;
+        this.alarmMessage = alarmMessage;
+        this.alarming = alarming;
+    }
+
+    public FastJsonLongIdKey getKey() {
+        return key;
+    }
+
+    public void setKey(FastJsonLongIdKey key) {
+        this.key = key;
+    }
+
+    public long getLimitCapacity() {
+        return limitCapacity;
+    }
+
+    public void setLimitCapacity(long limitCapacity) {
+        this.limitCapacity = limitCapacity;
+    }
+
+    public long getActualCapacity() {
+        return actualCapacity;
+    }
+
+    public void setActualCapacity(long actualCapacity) {
+        this.actualCapacity = actualCapacity;
+    }
+
+    public double getRatio() {
+        return ratio;
+    }
+
+    public void setRatio(double ratio) {
+        this.ratio = ratio;
+    }
+
+    public Date getHappenedDate() {
+        return happenedDate;
+    }
+
+    public void setHappenedDate(Date happenedDate) {
+        this.happenedDate = happenedDate;
+    }
+
+    public String getAlarmMessage() {
+        return alarmMessage;
+    }
+
+    public void setAlarmMessage(String alarmMessage) {
+        this.alarmMessage = alarmMessage;
+    }
+
+    public boolean isAlarming() {
+        return alarming;
+    }
+
+    public void setAlarming(boolean alarming) {
+        this.alarming = alarming;
+    }
+
+    @Override
+    public String toString() {
+        return "FastJsonAlarmInfo{" +
+                "key=" + key +
+                ", limitCapacity=" + limitCapacity +
+                ", actualCapacity=" + actualCapacity +
+                ", ratio=" + ratio +
+                ", happenedDate=" + happenedDate +
+                ", alarmMessage='" + alarmMessage + '\'' +
+                ", alarming=" + alarming +
+                '}';
+    }
+}