+++ capacity-checker-stack/src/main/java/com/dwarfeng/capacitychecker/stack/dao/AlarmInfoDao.java
@@ -1,0 +1,17 @@
+package com.dwarfeng.capacitychecker.stack.dao;
+
+import com.dwarfeng.capacitychecker.stack.bean.entity.AlarmInfo;
+import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
+import com.dwarfeng.subgrade.stack.dao.BatchBaseDao;
+import com.dwarfeng.subgrade.stack.dao.EntireLookupDao;
+import com.dwarfeng.subgrade.stack.dao.PresetLookupDao;
+
+/**
+ * 报警信息数据访问层。
+ *
+ * @author DwArFeng
+ * @since 1.0.0
+ */
+public interface AlarmInfoDao extends BatchBaseDao<LongIdKey, AlarmInfo>, EntireLookupDao<AlarmInfo>,
+        PresetLookupDao<AlarmInfo> {
+}